Another iteration at templating the thermo directory to get Cantera to return Jacobians. This still doesn't compile.

+++ cantera/branches/jacobians/src/thermo/SpeciesThermoInterpType.cpp
@@ -36,7 +36,7 @@ doublereal SpeciesThermoInterpType::reportHf298(doublereal* const h298) const
                        "Not implemented");
 }
 
-void SpeciesThermoInterpType::modifyOneHf298(const int k, const doublereal Hf298New)
+void SpeciesThermoInterpType::modifyOneHf298(const size_t k, const doublereal Hf298New)
 {
     throw CanteraError("SpeciesThermoInterpType::modifyOneHf298",
                        "Not implemented");
